Added examples + refactoring of e2e tests
