+++ .github/workflows/tests.yml
@@ -66,7 +66,7 @@ jobs:
       - name: Wait database available
         run: bash ./.github/scripts/wait-ydb-container.sh
       - name: Test
-        run: go test -race -coverpkg=./... -coverprofile table.txt -covermode atomic ./table_test.go
+        run: go test -race -coverpkg=./... -coverprofile table.txt -covermode atomic ./table/table_test.go
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2
         with:
@@ -107,7 +107,7 @@ jobs:
       - name: Wait database available
         run: bash ./.github/scripts/wait-ydb-container.sh
       - name: Test
-        run: go test -race -coverpkg=./... -coverprofile ratelimiter.txt -covermode atomic ./ratelimiter_test.go
+        run: go test -race -coverpkg=./... -coverprofile ratelimiter.txt -covermode atomic ./ratelimiter/ratelimiter_test.go
         shell: bash
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2
@@ -149,7 +149,7 @@ jobs:
       - name: Wait database available
         run: bash ./.github/scripts/wait-ydb-container.sh
       - name: Test
-        run: go test -race -coverpkg=./... -coverprofile scripting.txt -covermode atomic ./scripting_test.go
+        run: go test -race -coverpkg=./... -coverprofile scripting.txt -covermode atomic ./scripting/scripting_test.go
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2
         with:
@@ -190,7 +190,7 @@ jobs:
       - name: Wait database available
         run: bash ./.github/scripts/wait-ydb-container.sh
       - name: Test
-        run: go test -race -coverpkg=./... -coverprofile discovery.txt -covermode atomic ./discovery_test.go
+        run: go test -race -coverpkg=./... -coverprofile discovery.txt -covermode atomic ./discovery/discovery_test.go
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2
         with:
@@ -238,44 +238,3 @@ jobs:
           file: ./connection.txt
           flags: connection,e2e,integration,${{ matrix.os }},${{ matrix.go-version }}
           name: connection
-  long-stream:
-    strategy:
-      matrix:
-        go-version: [1.17.x]
-        os: [ubuntu-latest]
-    services:
-      ydb:
-        image: cr.yandex/yc/yandex-docker-local-ydb:latest
-        ports:
-          - 2135:2135
-          - 2136:2136
-          - 8765:8765
-        volumes:
-          - /tmp/ydb_certs:/ydb_certs
-        env:
-          YDB_LOCAL_SURVIVE_RESTART: true
-          YDB_USE_IN_MEMORY_PDISKS: true
-        options: '-h localhost'
-    env:
-      OS: ${{ matrix.os }}
-      GO: ${{ matrix.go-version }}
-      YDB_CONNECTION_STRING: grpcs://localhost:2135/?database=/local
-      YDB_SSL_ROOT_CERTIFICATES_FILE: /tmp/ydb_certs/ca.pem
-    runs-on: ${{ matrix.os }}
-    steps:
-      - name: Install Go
-        uses: actions/setup-go@v2
-        with:
-          go-version: ${{ matrix.go-version }}
-      - name: Checkout code
-        uses: actions/checkout@v2
-      - name: Wait database available
-        run: bash ./.github/scripts/wait-ydb-container.sh
-      - name: Test
-        run: go test -race -coverpkg=./... -coverprofile long-stream.txt -covermode atomic ./long_stream_test.go
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v2
-        with:
-          file: ./long-stream.txt
-          flags: long-stream,e2e,integration,${{ matrix.os }},${{ matrix.go-version }}
-          name: long-stream
