+++ discovery/discovery_test.go
@@ -1,11 +1,12 @@
 //go:build !fast
 // +build !fast
 
-package ydb_test
+package discovery_test
 
 import (
 	"context"
 	"crypto/tls"
+	"fmt"
 	"os"
 	"testing"
 	"time"
@@ -125,4 +126,39 @@ func TestDiscovery(t *testing.T) {
 			t.Fatalf("Execute failed: %v", err)
 		}
 	})
+}
+
+func Example_discoverCluster() {
+	ctx := context.TODO()
+	db, err := ydb.Open(ctx, "grpcs://localhost:2135/?database=/local")
+	if err != nil {
+		fmt.Printf("failed to connect: %v", err)
+		return
+	}
+	defer db.Close(ctx) // cleanup resources
+	endpoints, err := db.Discovery().Discover(ctx)
+	if err != nil {
+		fmt.Printf("discover failed: %v", err)
+		return
+	}
+	fmt.Printf("%s endpoints:\n", db.Name())
+	for i, e := range endpoints {
+		fmt.Printf("%d) %s\n", i, e.String())
+	}
+}
+
+func Example_whoAmI() {
+	ctx := context.TODO()
+	db, err := ydb.Open(ctx, "grpcs://localhost:2135/?database=/local")
+	if err != nil {
+		fmt.Printf("failed to connect: %v", err)
+		return
+	}
+	defer db.Close(ctx) // cleanup resources
+	whoAmI, err := db.Discovery().WhoAmI(ctx)
+	if err != nil {
+		fmt.Printf("discover failed: %v", err)
+		return
+	}
+	fmt.Printf("%s whoAmI: %s\n", db.Name(), whoAmI.String())
 }
